chore: extend incremental filename enforcement to api-client giphy and sleep handling
