+++ libraries/api-client/src/giphy/giphyImage.ts
