+++ libraries/api-client/src/utils/backFromSleepHandler/index.ts
@@ -1,6 +1,6 @@
 /*
  * Wire
- * Copyright (C) 2018 Wire Swiss GmbH
+ * Copyright (C) 2024 Wire Swiss GmbH
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -17,22 +17,4 @@
  *
  */
 
-import {GiphyImage} from './GiphyImage';
-
-export interface GiphyResult {
-  data: GiphyImage;
-  meta: {
-    msg: string;
-    response_id: string;
-    status: number;
-  };
-}
-
-export type GiphyMultipleResult = GiphyResult & {
-  data: GiphyImage[];
-  pagination: {
-    count: number;
-    offset: number;
-    total_count: number;
-  };
-};
+export * from './backFromSleepHandler';
